+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/utils/__init__.py
@@ -17,7 +17,7 @@ Sub-module containing utilities for the Orbit framework.
 from .array import TENSOR_TYPE_CONVERSIONS, TENSOR_TYPES, TensorData, convert_to_torch
 from .configclass import configclass
 from .dict import class_to_dict, convert_dict_to_backend, print_dict, update_class_from_dict, update_dict
-from .string import to_camel_case, to_snake_case
+from .string import callable_to_string, is_lambda_expression, string_to_callable, to_camel_case, to_snake_case
 from .timer import Timer
 
 __all__ = [
@@ -37,6 +37,9 @@ __all__ = [
     # string utilities
     "to_camel_case",
     "to_snake_case",
+    "is_lambda_expression",
+    "string_to_callable",
+    "callable_to_string",
     # timer
     "Timer",
 ]