+++ source/extensions/omni.isaac.orbit/test/test_dict_utils.py
@@ -8,6 +8,16 @@ import unittest
 import omni.isaac.orbit.utils.dict as dict_utils
 
 
+def test_function(x):
+    """Test function for string <-> callable conversion."""
+    return x**2
+
+
+def test_lambda_function(x):
+    """Test function for string <-> callable conversion."""
+    return x**2
+
+
 class TestDictUtilities(unittest.TestCase):
     """Test fixture for checking Kit utilities in Orbit."""
 
@@ -24,6 +34,38 @@ class TestDictUtilities(unittest.TestCase):
         }
         # print the dictionary
         dict_utils.print_dict(test_dict)
+
+    def test_string_callable_function_conversion(self):
+        """Test string <-> callable conversion for function."""
+
+        # convert function to string
+        test_string = dict_utils.callable_to_string(test_function)
+        # convert string to function
+        test_function_2 = dict_utils.string_to_callable(test_string)
+        # check that functions are the same
+        self.assertEqual(test_function(2), test_function_2(2))
+
+    def test_string_callable_function_with_lambda_in_name_conversion(self):
+        """Test string <-> callable conversion for function which has lambda in its name."""
+
+        # convert function to string
+        test_string = dict_utils.callable_to_string(test_lambda_function)
+        # convert string to function
+        test_function_2 = dict_utils.string_to_callable(test_string)
+        # check that functions are the same
+        self.assertEqual(test_function(2), test_function_2(2))
+
+    def test_string_callable_lambda_conversion(self):
+        """Test string <-> callable conversion for lambda expression."""
+
+        # create lambda function
+        func = lambda x: x**2
+        # convert function to string
+        test_string = dict_utils.callable_to_string(func)
+        # convert string to function
+        func_2 = dict_utils.string_to_callable(test_string)
+        # check that functions are the same
+        self.assertEqual(func(2), func_2(2))
 
 
 if __name__ == "__main__":